Add selector and refactor container
